[routes]organise Search, UserPosts and Users routes
