+++ backend/src/routes/Search/getAllUsers.ts
@@ -1,0 +1,15 @@
+import { Request, Response, NextFunction } from "express";
+import User from "../../models/User.models";
+
+export const getAllUsers = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
+  try {
+    const allUsers = await User.find();
+    res.status(200).json(allUsers);
+  } catch (error) {
+    next(error);
+  }
+};
